requirements.txt
# ============================================================
# LLM Council MGA — Python Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# Core framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
starlette>=0.36.0

# HTTP client (async, for LLM API calls)
httpx>=0.27.0,<1.0.0

# Data validation
pydantic>=2.9.0,<3.0.0

# Environment config
python-dotenv>=1.0.0,<2.0.0

# ── Testing (optional) ──────────────────────────────────────
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
